rename internal/search/commits to internal/search/commit
